requirements.txt

streamlit==1.37.1
PyPDF2==3.0.1
python-docx==1.1.2
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
